+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 ## 1.0.0-beta.4
 
+* Fix a crash when `:not(...)` extends a selector that appears in
+  `:not(:not(...))`.
+
+### Node JS API
+
+* Add support for asynchronous importers to `render()` and `renderSync()`.
+
+### Dart API
+
 * Add `compileAsync()` and `compileStringAsync()` methods. These run
   asynchronously, which allows them to take asynchronous importers (see below).
 
 * Add an `AsyncImporter` class. This allows imports to be resolved
   asynchronously in case no synchronous APIs are available. `AsyncImporter`s are
   only compatible with `compileAysnc()` and `compileStringAsync()`.
-
-* Fix a crash when `:not(...)` extends a selector that appears in
-  `:not(:not(...))`.
 
 ## 1.0.0-beta.3